"Fixed" the invisible dialog error, as in #209
The problem was that if you center a dialog wider/taller than the screen, it will overflow. Since albow draws widgets by openGL, it won't accept negative coordinates for the bottom left corner. my hotfix minimizes these coordinates.

This is not a complete solution, and some of the filters, for example, will still be out of the screen. At least now we see what's going on.